requirements.txt
requests
python-dotenv
graphqlclient